e2e/requirements.txt
pytest==7.2.1
pytest-cov==4.1.0
pytest-asyncio==0.21.0
parameterized==0.8.1
telethon==1.28.5
python-telegram-bot[webhooks]==20.3
